refactoring proyect structure
