+++ catalogo-arqueologico/src/pages/ObjectDetail/components/PieceVisualization.jsx
